// extractors/backend/clients/messages/Client.ts
import { z } from "zod"

export const clientValidator = z.object({
    member_id: z.number(),
    unique_id: z.number(),
    record_type: z.string(),
    email: z.string(),
    full_name: z.string(),
    registration_date: z.date(),
    last_login_date: z.date(),
    last_active_date: z.date(),
    cookie_path:z.string(),
    metadata: z.object({
      isMobile: z.boolean().nullable().optional(),
      leadAssignmentJob: z.number().nullable().optional(),
      autogeneratedPassword: z.boolean().nullable().optional(),
    }).nullable(),
    created_at: z.date(),
    updated_at: z.date(),
    deleted_at: z.date().nullable(),
    first_name: z.string(),
    last_name: z.string(),
    promo_code: z.string().nullable(),
    profile_information: z.object({
      country: z.string().nullable().optional(),
    }).nullable(),
    additional_attributes: z.object({
      qftd: z.boolean().nullable(),
      locale: z.string().nullable(),
      opt_in: z.boolean().nullable(),
      company:  z.string().nullable(),
      is_pool: z.boolean().nullable(),
      nickname:  z.string().nullable(),
      campaign_id:  z.string().nullable(),
      client_type:  z.string().nullable(),
      cookie_path: z.unknown().nullable(),
      white_label: z.string().nullable(),
      affiliate_id: z.string().nullable(),
      needs_verify: z.number().nullable(),
      member_p_param: z.string().nullable(),
      sales_agent_id: z.number().nullable(),
      trading_status: z.string().nullable(),
      support_agent_id: z.number().nullable(),
      conversion_status: z.string().nullable(),
      sales_agent_email: z.string().nullable(),
      member_campaign_id:z.string().nullable(),
      support_agent_email: z.string().nullable(),
      trading_status_reason: z.string().nullable(),
      trading_status_reason_code: z.string().nullable(),
    }).nullable(),
    kyc_status: z.string(),
    converted_at: z.date().nullable(),
    kyc_rejection_reason_code: z.string().nullable(),
    query_submission_timestamp: z.number(),
    update_epoch: z.number(),
})
export type ClientValidator = typeof clientValidator
export type Client = z.infer<typeof clientValidator>